Add a link to Google Sky's web edition.
For now, this is only a simple link to the site. It's missing
"longitude" and "latitude" parameters (-> RA/Dec) for now and
the KMZ file cannot be given in the URL either. However, this
is work in progress: I asked internal Google staff for this
to be added. Hope to get a positive answer!

git-svn-id: svn+ssh://astrometry.net/svn/trunk/src/astrometry@19988 03a3cea6-2c03-0410-ac9b-9271b1c66c29